➕ Add link to reqwest-client feature (#139)
The opentelmetry-otlp package by default loads the opentelemetry-http package in such a way that
it does not have an actual http client available, which causes it to error on initialization. Add
in a link to the opentelmetry-otlp reqwest-client feature, which will in turn load the reqwest
crate and link in the opentelemetry-http crate's reqwest feature.